Added search based on name, added query like search for financial values.
still working to add more, basic functionality is done for these features. there may not be much more to the test, but would like to keep going and do more if i can with this and the asp.net. there is a slight bug with the searches, they are only working independantly of one another right now, would like to see that fixed... otherwise everything is perfect exactly what i wanted...