UI Overhaul and Added autoscrolling and flash animation for edit
this was some Quality of life edits. did a bunch of front end work. added bootstrap and overhauled the ui to be more modern, include the cmu logo and use its color scheme that i found on their website. as well as added return scrolling and a flash animation to show the edited field when an edit has been completed. will have more details in my notes about the changes here.